Move Colonini parser types into new files
Does not move the functions into these files yet.
